+++ MSFileToolbox/src/umich/ms/fileio/filetypes/pepxml/jaxb/primitive/InterprophetSummary.java
@@ -43,8 +43,8 @@ import javax.xml.bind.annotation.XmlType;
  *       &lt;/sequence>
  *       &lt;attribute name="version" use="required" type="{http://www.w3.org/2001/XMLSchema}string" />
  *       &lt;attribute name="options" type="{http://www.w3.org/2001/XMLSchema}string" />
- *       &lt;attribute name="est_tot_num_correct_psm" type="{http://www.w3.org/2001/XMLSchema}float" />
- *       &lt;attribute name="est_tot_num_correct_pep" type="{http://www.w3.org/2001/XMLSchema}float" />
+ *       &lt;attribute name="est_tot_num_correct_psm" type="{http://www.w3.org/2001/XMLSchema}double" />
+ *       &lt;attribute name="est_tot_num_correct_pep" type="{http://www.w3.org/2001/XMLSchema}double" />
  *     &lt;/restriction>
  *   &lt;/complexContent>
  * &lt;/complexType>
@@ -71,9 +71,9 @@ public class InterprophetSummary {
     @XmlAttribute(name = "options")
     protected String options;
     @XmlAttribute(name = "est_tot_num_correct_psm")
-    protected float estTotNumCorrectPsm;
+    protected double estTotNumCorrectPsm;
     @XmlAttribute(name = "est_tot_num_correct_pep")
-    protected float estTotNumCorrectPep;
+    protected double estTotNumCorrectPep;
 
     /**
      * Gets the value of the inputfile property.
@@ -214,7 +214,7 @@ public class InterprophetSummary {
      * Gets the value of the estTotNumCorrectPsm property.
      * 
      */
-    public float getEstTotNumCorrectPsm() {
+    public double getEstTotNumCorrectPsm() {
         return estTotNumCorrectPsm;
     }
 
@@ -222,7 +222,7 @@ public class InterprophetSummary {
      * Sets the value of the estTotNumCorrectPsm property.
      * 
      */
-    public void setEstTotNumCorrectPsm(float value) {
+    public void setEstTotNumCorrectPsm(double value) {
         this.estTotNumCorrectPsm = value;
     }
 
@@ -230,7 +230,7 @@ public class InterprophetSummary {
      * Gets the value of the estTotNumCorrectPep property.
      * 
      */
-    public float getEstTotNumCorrectPep() {
+    public double getEstTotNumCorrectPep() {
         return estTotNumCorrectPep;
     }
 
@@ -238,7 +238,7 @@ public class InterprophetSummary {
      * Sets the value of the estTotNumCorrectPep property.
      * 
      */
-    public void setEstTotNumCorrectPep(float value) {
+    public void setEstTotNumCorrectPep(double value) {
         this.estTotNumCorrectPep = value;
     }
